redo nativity parsing algorithm
Completely redo the algorithm which parses nativities from text and create a class for it in the 'Api' assembly so it can be used by plugins also.

The new algorithm does not count numerically and knows how to recognize id numbers and titles correctly. It is also easier to update the algorithm so it knows how to detect separate ids from titles better.